Extract reservation listing tests to own file.
